Ulti_argus/To_Recycle/Argus_V_clone/Argus_V-main/requirements.txt
# ARGUS_V pinned production dependencies
#
# These pins are intended to work on Raspberry Pi OS (Debian-based).
# If you are using an older Raspberry Pi OS image, you may need to adjust
# numpy/scipy/scikit-learn pins based on available wheels.

PyYAML==6.0.2
scapy==2.6.1
# pcapy-ng==1.0.9
pytz==2025.2
firebase-admin==6.6.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
skops==0.13.0
